src/Retrain_Consumer/requirements.txt
pandas
kafka-python
redis
joblib
